fix(bdd): move story-a helpers under rootDir to fix TS6059 build break
Task 6 (d8eb7d09) extracted shared helpers to tests/bdd/support/helpers.ts,

but story-a-acceptance.test.ts in src/__tests__/ imports them. Since the

acceptance test is part of the tsc build program (include: src/**/*), importing

a file outside rootDir (./src) triggers TS6059, breaking npm run build and

verify:merge gate 7.

Fix: move helpers.ts to src/runtime-v2/activation/__tests__/helpers.ts

(under rootDir). Register it in io-seam-registry.json under a new

'test-infrastructure' seam because it does real I/O (mkdtempSync, rmSync,

SQLite connection) and core's no-restricted-imports rule bans fs/path in src/.

Both ESLint and architecture-regression.test.ts derive exemptions from the

registry, so both guards update automatically.

Updated import paths in:

- story-a-acceptance.test.ts: ./helpers.js (same dir)

- tests/bdd/story-a.steps.ts: ../../src/runtime-v2/activation/__tests__/helpers.js
