+++ packages/principles-core/io-seam-registry.json
@@ -61,6 +61,13 @@
         "runtime-v2/internalization-integrity-remediation.ts",
         "runtime-v2/pruning-review-log.ts"
       ]
+    },
+    {
+      "name": "test-infrastructure",
+      "description": "Shared Story A acceptance test helpers that must live under src/ (rootDir) because story-a-acceptance.test.ts in src/__tests__/ imports them — placing them in tests/ triggers TS6059 (File is not under rootDir). Does I/O for temp workspace setup (mkdtempSync, rmSync) and SQLite connection initialization. NOT production code; lives in src/ solely due to the tsc rootDir constraint. BDD steps (tests/bdd/story-a.steps.ts) cross-import these helpers.",
+      "files": [
+        "runtime-v2/activation/__tests__/helpers.ts"
+      ]
     }
   ]
 }
